Move commands to a separate module
